todo_backend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.12.0
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
